WIP: generic_drug works, now tests are needed
